Implement leaving groups/deleting groups. Restructure account endpoints

+++ cliqcard_server/views/account.py
@@ -1,6 +1,5 @@
 import time
-from flask import request, jsonify
-from flask.views import MethodView
+from flask import request, jsonify, Blueprint
 from authlib.flask.oauth2 import current_token
 from authlib.common.security import generate_token
 from cliqcard_server.models import db, User, RegistrationToken, Card, OAuthClient, OAuthToken
@@ -10,110 +9,15 @@ from cliqcard_server.extensions import bcrypt
 from cliqcard_server.errors import InvalidRequestError, UnauthorizedError
 
 
-class AccountView(MethodView):
-    """
-    Handles the /account endpoint for getting, updating, and deleting accounts.
-    """
+account = Blueprint('account', __name__, url_prefix='/account')
 
-    @require_oauth(None)
-    def get(self):
+
+@account.route('/', methods=['GET', 'PUT', 'DELETE'])
+@require_oauth(None)
+def index():
+    if request.method == 'GET':
         return jsonify(serialize_account(current_token.user))
-
-
-    def post(self):
-        # get client credentials
-        client_id = request.json.get('client_id')
-        client_secret = request.json.get('client_secret')
-
-        # find the client
-        client = OAuthClient.query.filter_by(client_id=client_id, client_secret=client_secret).first()
-        if not client:
-            raise UnauthorizedError(message='Invalid client credentials')
-
-        # get phone number and registration token
-        phone_number = request.json.get('phone_number')
-        raw_registration_token = request.json.get('registration_token')
-
-        # get other required user details
-        first_name = request.json.get('first_name')
-        last_name = request.json.get('last_name')
-
-        if first_name is None:
-            raise InvalidRequestError('Missing \'first_name\' parameter')
-        if last_name is None:
-            raise InvalidRequestError('Missing \'last_name\' parameter')
-
-        # get optional user details
-        email = request.json.get('email', None)
-
-        # validate/format phone number
-        phone_number = format_phone_number(phone_number)
-        if not phone_number:
-            raise InvalidRequestError('Invalid phone number. You must send phone numbers in valid international format')
-
-        # find the matching registration token for this phone number
-        registration_token = RegistrationToken.query.filter_by(phone_number=phone_number).first()
-        if not registration_token:
-            raise UnauthorizedError()
-        elif registration_token.is_expired():
-            # token is expired - delete from database
-            db.session.delete(registration_token)
-            db.session.commit()
-            raise InvalidRequestError('The registration code for that phone number has expired')
-
-        # check if the tokens match
-        if not bcrypt.check_password_hash(registration_token.token, raw_registration_token):
-            # no match - wrong token
-            raise UnauthorizedError()
-
-        # delete the registration token
-        db.session.delete(registration_token)
-
-        # create a new user
-        user = User(
-            phone_number=phone_number,
-            first_name=first_name,
-            last_name=last_name,
-            email=email
-        )
-        db.session.add(user)
-
-        # commit the session - important because we need the user id
-        db.session.commit()
-
-        # create personal and work cards for the user
-        user.personal_card = Card(phone1=phone_number, email=email)
-        user.work_card = Card()
-
-        # generate an access token for the user
-        token = OAuthToken()
-        token.client_id = client.client_id,
-        token.token_type = 'Bearer',
-        token.access_token = generate_token(),
-        token.refresh_token = generate_token(),
-        token.issued_at = int(time.time()),
-        token.expires_in = 3600,
-        token.user_id = user.id
-        db.session.add(token)
-
-        # commit the session
-        db.session.commit()
-
-        # respond with the newly created account info
-        response = jsonify({
-            'token': {
-                'access_token': token.access_token,
-                'refresh_token': token.refresh_token,
-                'token_type': token.token_type,
-                'expires_in': token.expires_in
-            },
-            'user': serialize_account(user)
-        })
-        response.status_code = 201
-        return response
-
-    @require_oauth(None)
-    def put(self):
+    elif request.method == 'PUT':
         user = current_token.user
 
         phone_number = request.json.get('phone_number')
@@ -121,7 +25,8 @@ class AccountView(MethodView):
             # validate/format phone number
             phone_number = format_phone_number(phone_number)
             if not phone_number:
-                raise InvalidRequestError('Invalid phone number. You must send phone numbers in valid international format')
+                raise InvalidRequestError(
+                    'Invalid phone number. You must send phone numbers in valid international format')
 
         user.phone_number = phone_number if phone_number is not None else user.phone_number
         user.first_name = request.json.get('first_name', user.first_name)
@@ -131,9 +36,101 @@ class AccountView(MethodView):
         db.session.commit()
 
         return jsonify(serialize_account(user))
-
-    @require_oauth(None)
-    def delete(self):
+    elif request.method == 'DELETE':
         db.session.delete(current_token.user)
         db.session.commit()
         return ('', 204)
+
+
+@account.route('/', methods=['POST'])
+def post():
+    # get client credentials
+    client_id = request.json.get('client_id')
+    client_secret = request.json.get('client_secret')
+
+    # find the client
+    client = OAuthClient.query.filter_by(client_id=client_id, client_secret=client_secret).first()
+    if not client:
+        raise UnauthorizedError(message='Invalid client credentials')
+
+    # get phone number and registration token
+    phone_number = request.json.get('phone_number')
+    raw_registration_token = request.json.get('registration_token')
+
+    # get other required user details
+    first_name = request.json.get('first_name')
+    last_name = request.json.get('last_name')
+
+    if first_name is None:
+        raise InvalidRequestError('Missing \'first_name\' parameter')
+    if last_name is None:
+        raise InvalidRequestError('Missing \'last_name\' parameter')
+
+    # get optional user details
+    email = request.json.get('email', None)
+
+    # validate/format phone number
+    phone_number = format_phone_number(phone_number)
+    if not phone_number:
+        raise InvalidRequestError('Invalid phone number. You must send phone numbers in valid international format')
+
+    # find the matching registration token for this phone number
+    registration_token = RegistrationToken.query.filter_by(phone_number=phone_number).first()
+    if not registration_token:
+        raise UnauthorizedError()
+    elif registration_token.is_expired():
+        # token is expired - delete from database
+        db.session.delete(registration_token)
+        db.session.commit()
+        raise InvalidRequestError('The registration code for that phone number has expired')
+
+    # check if the tokens match
+    if not bcrypt.check_password_hash(registration_token.token, raw_registration_token):
+        # no match - wrong token
+        raise UnauthorizedError()
+
+    # delete the registration token
+    db.session.delete(registration_token)
+
+    # create a new user
+    user = User(
+        phone_number=phone_number,
+        first_name=first_name,
+        last_name=last_name,
+        email=email
+    )
+    db.session.add(user)
+
+    # commit the session - important because we need the user id
+    db.session.commit()
+
+    # create personal and work cards for the user
+    user.personal_card = Card(phone1=phone_number, email=email)
+    user.work_card = Card()
+
+    # generate an access token for the user
+    token = OAuthToken()
+    token.client_id = client.client_id,
+    token.token_type = 'Bearer',
+    token.access_token = generate_token(),
+    token.refresh_token = generate_token(),
+    token.issued_at = int(time.time()),
+    token.expires_in = 3600,
+    token.user_id = user.id
+    db.session.add(token)
+
+    # commit the session
+    db.session.commit()
+
+    # respond with the newly created account info
+    response = jsonify({
+        'token': {
+            'access_token': token.access_token,
+            'refresh_token': token.refresh_token,
+            'token_type': token.token_type,
+            'expires_in': token.expires_in
+        },
+        'user': serialize_account(user)
+    })
+    response.status_code = 201
+    return response
